fix: create stories without histoire
