+++ desktop/package.json
@@ -16,9 +16,8 @@
 		"dev": "vite",
 		"build": "vite build",
 		"preview": "vite preview",
-		"story:dev": "histoire dev",
-		"story:build": "histoire build",
-		"story:preview": "histoire preview",
+		"story:dev": "vite dev stories/ -c stories/vite.config.ts",
+		"story:build": "vite build stories/ -c stories/vite.config.ts",
 		"lint": "eslint . --ext .ts,.vue"
 	},
 	"dependencies": {
@@ -31,7 +30,6 @@
 		"xstate": "~4.33.6"
 	},
 	"devDependencies": {
-		"@histoire/plugin-vue": "^0.13.0",
 		"@miragejs/graphql": "^0.1.13",
 		"@typescript-eslint/eslint-plugin": "^5.33.0",
 		"@typescript-eslint/parser": "^5.33.0",
@@ -39,7 +37,6 @@
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-vue": "^9.3.0",
 		"eslint": "^8.22.0",
-		"histoire": "^0.13.0",
 		"miragejs": "^0.1.45",
 		"typescript": "^4.9.4",
 		"vite": "^4.1.1",
