--- desktop/stories/Dev.vue
@@ -1,9 +1,0 @@
-<template>
-	<router-view v-slot="{ Component, route }">
-		<transition :name="route.meta.transition || 'slide-up'">
-			<component :is="Component" />
-		</transition>
-	</router-view>
-</template>
-
-<script setup lang="ts"></script>
