+++ desktop/stories/components/Fieldtype.vue
@@ -9,9 +9,10 @@ const app = getCurrentInstance()
 
 const registeredComponents = computed(() => {
 	let components = []
-	for (const [key, value] of Object.entries(app.appContext.components)) {
-		if (value.props.register) {
-			components.push(key) // may be appropriate to pass more than just the instance name here
+	for (const [name, obj] of Object.entries(app.appContext.components)) {
+		if (obj.props?.register) {
+			// may be appropriate to pass more than just the instance name here
+			components.push(name)
 		}
 	}
 	return components
