+++ examples/tsconfig.json
@@ -1,0 +1,11 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"composite": true,
+		"declarationMap": true,
+		"emitDeclarationOnly": true,
+		"outDir": "dist",
+		"baseUrl": "."
+	},
+	"include": ["docbuilder/**/*", "docbuilder/**/*.ts", "docbuilder/**/*.vue", "docbuilder/**/*.json"]
+}
